+++ spec/spec_helper.rb
@@ -1,15 +1,22 @@
 # frozen_string_literal: true
 
 require 'ncba'
+require 'webmock/rspec'
 
 RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = '.rspec_status'
-
-  # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
+  end
+
+  config.before do
+    Ncba.reset_configuration!
+    Ncba.configure do |c|
+      c.api_key = 'test_api_key'
+      c.username = 'test_user'
+      c.password = 'test_pass'
+    end
   end
 end